Fold the race and reset-token tests into their pair files
